% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_model_eq.R
\name{vis_model_eq}
\alias{vis_model_eq}
\title{Visualize the functional form of a CSM equation}
\usage{
vis_model_eq(model, name, mapping, ...)
}
\arguments{
\item{model}{a list vector containing a CSM as created by
[csmbuilder::csm_create_model()]}

\item{name}{a character string of the variable name whose equation should
 be visualized

 @param mapping an aesthetic mapping as defined by [ggplot2::aes()]}

\item{...}{other arguments corresponding to variables in the model equation
to be visualized}
}
\description{
Visualize the functional form of a CSM equation
}
